StockAnalysis/app/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
flask-caching
flask-migrate
celery
redis
python-dotenv
pandas
numpy
nselib
transformers